tidy up modules
