+++ Cargo.toml
@@ -12,6 +12,3 @@ members = [
 
 [profile.release]
 lto = "thin"
-
-[patch.crates-io]
-k8s-gateway-api = { git = "https://github.com/the-wondersmith/k8s-gateway-api-rs", branch = "feat-grpcroute-support" }
