+++ policy-controller/k8s/index/src/inbound/tests.rs
@@ -227,6 +227,7 @@ impl TestConfig {
                 timeout: self.detect_timeout,
             },
             http_routes: mk_default_routes(),
+            grpc_routes: Default::default(),
         }
     }
